+++ nuxeo-drive-client/nxdrive/data/ui5/js/ndrive-conflicts.js
@@ -46,6 +46,7 @@ ConflictsController.prototype.updateErrors = function($scope) {
 	        'LOCKED',
 	        'MANUAL',
 	        'DEDUP',
+	        'CORRUPT',
 	    ];
 
 	for (error in errors) {